working through cleanup
