+++ examples/julia/default.json
@@ -33,11 +33,11 @@
       "lookup_table_count": 2048,
       "background_color_rgb": [0, 0, 0],
       "histogram_bin_count": 32,
-      "histogram_sample_count": 50000
+      "histogram_sample_count": 40000
     },
     "render_options": {
       "downsample_stride": 1,
-      "subpixel_antialiasing": 3
+      "subpixel_antialiasing": 2
     }
   }
 }
