+++ src/core/render_quality_fsm.rs
@@ -1,12 +1,3 @@
-//! Simple FSM (finite state machine) that is used to regulate the
-//! render quality command for the render pipeline. While the user
-//! is actively interacting with the system, we want to hit a target
-//! frame rate, even if the render quality is low. However, once the
-//! user stops interacting, then we need to quickly crank up the render
-//! quality regardless of frame rate. Finally, once we've rendered at
-//! high quality, we should shut down the render pipeline to conserve
-//! resources (no need to spin at max CPU while idle...).
-
 pub trait RenderQualityPolicy {
     const MAX_COMMAND: f64 = 1.0;
     const MIN_COMMAND: f64 = 0.0;
@@ -51,6 +42,14 @@ pub enum Mode {
     Idle,
 }
 
+/// Simple FSM (finite state machine) that is used to regulate the
+/// render quality command for the render pipeline. While the user
+/// is actively interacting with the system, we want to hit a target
+/// frame rate, even if the render quality is low. However, once the
+/// user stops interacting, then we need to quickly crank up the render
+/// quality regardless of frame rate. Finally, once we've rendered at
+/// high quality, we should shut down the render pipeline to conserve
+/// resources (no need to spin at max CPU while idle...).
 #[derive(Debug, Clone)]
 pub struct FiniteStateMachine<F, G>
 where
@@ -163,6 +162,92 @@ where
             Some(self.previous_interactive_render_command)
         } else {
             None
+        }
+    }
+}
+
+////////////////////////////////////////////////////////////////////////////////////////
+
+/// The `AdaptiveOptimizationRegulator` is a simple class wrapping a finite state machine
+/// that is used to compute the "render quality" (0 = high quality but slow, 1 = low quality
+/// but fast), while exploring a fractal interactively with the user.
+#[derive(Clone, Debug)]
+pub struct AdaptiveOptimizationRegulator {
+    render_policy_fsm: FiniteStateMachine<ConstantFrameRatePolicy, ConstantFrameRatePolicy>,
+    render_start_time: Option<f64>,
+    render_period: Option<f64>,
+    render_command: Option<f64>,
+}
+
+/// For now, keep the regulator simple with some hard-coded policies.
+/// Eventually these will be replaced with policies that depend on the
+/// measured frame rate data.
+impl Default for AdaptiveOptimizationRegulator {
+    fn default() -> Self {
+        Self {
+            render_policy_fsm: FiniteStateMachine::new(
+                0.0,
+                ConstantFrameRatePolicy::new(0.55),
+                ConstantFrameRatePolicy::new(0.0),
+            ),
+            render_start_time: None,
+            render_period: None,
+            render_command: None,
+        }
+    }
+}
+
+impl AdaptiveOptimizationRegulator {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn reset(&mut self) {
+        self.render_policy_fsm.reset();
+        self.render_start_time = None;
+        self.render_period = None;
+        self.render_command = None;
+    }
+
+    /// This method is called each time that the `explore` pipeline would like
+    /// to render the fractal. It returns an optional value, which, if set,
+    /// indicates that the fractal should be rendered, and the floating point
+    /// value specifies the render quality value. If unset, in indicates that the
+    /// fractal already has been rendered to the screen, and does not need to be
+    /// recomputed, allowing the system to save resources.
+    pub fn render_required(&mut self, is_interactive: bool) -> Option<f64> {
+        self.render_policy_fsm.render_required(
+            self.render_command,
+            self.render_period,
+            is_interactive,
+        )
+    }
+
+    /// Called by the render pipeline whenever a new render begins.
+    /// This is a separate method from `render_required` because we cannot
+    /// run two renders at once, and the rendering happens in a separate
+    /// background process. This method will be called immediately at the
+    /// start of each enw render, and is used to collect accurate timing
+    /// data for the finite state machine logic. It caches that data for
+    /// use in the `render_required` method.
+    pub fn begin_rendering(&mut self, time: f64, command: f64) {
+        self.render_start_time = Some(time);
+        self.render_period = None;
+        self.render_command = Some(command);
+    }
+
+    /// Called by the render pipeline whenever the render is completed.
+    /// This is the matched method to the `begin_rendering` method, and
+    /// is used for accurate data collection on the frame rate. This method
+    /// should be called whenever the background thread finishes a render.
+    pub fn finish_rendering(&mut self, time: f64) {
+        // Note: this method will sometimes be called twice for a single
+        // `begin_rendering`, so we add this guard while will only update
+        // the period on the first call to `finish_rendering()` after calling
+        // `begin_rendering()`.
+        if let Some(start_time) = self.render_start_time {
+            self.render_period = Some(time - start_time);
+            self.render_start_time = None;
         }
     }
 }
